correction to countries phone int type

+++ database/migrations/2025_01_02_014855_create_countries_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('countries', function (Blueprint $table) {
-            $table->tinyInteger("code")->unsigned()->primary();
+            $table->smallInteger("code")->unsigned()->primary();
             $table->string("name");
             $table->tinyInteger("phone_length")->unsigned();
             $table->timestamps();
